[Broker] Change candidateIt to data. Fixed port type to int

+++ broker/src/main/java/broker/entities/Service.java
@@ -12,7 +12,7 @@ package broker.entities;
 public class Service {
     
     private String ip;
-    private long port;
+    private int port;
     private String service;
     
     public Service(String ip, int port, String services){
@@ -29,11 +29,11 @@ public class Service {
         this.ip = ip;
     }
 
-    public long getPort() {
+    public int getPort() {
         return port;
     }
 
-    public void setPort(long port) {
+    public void setPort(int port) {
         this.port = port;
     }
